refactor(rotation): extract the rotation engine + loop to internal/rotation (PR-D2, decouple prep)
The console-decouple's `watch` daemon (PR-D3) keeps `up`'s built-in unattended
rotation (#420), so the rotation engine — previously package-main in rotate.go —
must live in internal/ for a second binary to reach it. Core `rotate` and `up`
keep using it; behavior is unchanged.

internal/rotation:
- rotation.go: Perform(ctx, db, dbName, Options) (the old performRotation) + its
  partition helpers. The rot* flag globals it read directly become Options
  fields (slice-C-style); the engine no longer reads any global. HiveArchivePath
  is exported (the one cross-package use — archive/reconstruct tests build
  expected paths with it); every other helper stays private.
- loop.go: Settings/ParseSettings/StartLoop (the old upRotationSettings/
  parseUpRotation/startUpRotation) + the supervised-cycle helpers. The built-in
  loop now builds its Options via a pure loopOptions() (ProtectUnarchived armed,
  never archives) instead of fanning out to rot* globals — so the data-loss
  guard stays unit-tested (TestLoopOptions replaces the obsolete arming test).

cmd/bintrail:
- rotate.go gutted to the command + runRotate (builds Options → rotation.Perform);
  the rotProtectUnarchived global is gone (it only ever carried loop→engine state).
- up.go uses rotation.Settings / ParseSettings / StartLoop; up_rotation.go deleted.
- Tests of private helpers moved into package rotation; command-level tests stay.

internal/testutil.SetupPartitionedTable: the shared partition-setup helper the
rotation/archive/reconstruct integration tests all use, made dependency-free
(inline p_2006010215, no indexer import) so package-rotation tests can import
testutil without an import cycle.

Verified: build, vet (unit + integration), full unit suite, the relocated
rotation + cmd archive/reconstruct/monitor/up-rotation integration tests against
MySQL, and the E2E pipeline — all green. goimports scoped to the touched dirs.
