+++ cmd/bintrail/cmd_integration_test.go
@@ -270,140 +270,6 @@ func TestEnsureDatabase(t *testing.T) {
 	}
 }
 
-// ─── listPartitions ──────────────────────────────────────────────────────────────────
-
-func TestListPartitions(t *testing.T) {
-	db, dbName := testutil.CreateTestDB(t)
-
-	if err := indexer.CreateIndexTables(context.Background(), db, 3, false, nil); err != nil {
-		t.Fatalf("indexer.CreateIndexTables failed: %v", err)
-	}
-
-	parts, err := listPartitions(context.Background(), db, dbName)
-	if err != nil {
-		t.Fatalf("listPartitions failed: %v", err)
-	}
-
-	if len(parts) != 4 {
-		t.Fatalf("expected 4 partitions, got %d", len(parts))
-	}
-
-	// Verify ordinals are sequential.
-	for i, p := range parts {
-		if p.Ordinal != i+1 {
-			t.Errorf("partition %d: expected ordinal %d, got %d", i, i+1, p.Ordinal)
-		}
-	}
-
-	// Verify the last is p_future.
-	if parts[3].Name != "p_future" {
-		t.Errorf("expected last partition p_future, got %s", parts[3].Name)
-	}
-}
-
-// ─── dropPartitions ──────────────────────────────────────────────────────────────────
-
-func TestDropPartitions(t *testing.T) {
-	db, dbName := testutil.CreateTestDB(t)
-
-	if err := indexer.CreateIndexTables(context.Background(), db, 5, false, nil); err != nil {
-		t.Fatalf("indexer.CreateIndexTables failed: %v", err)
-	}
-
-	// List the first partition to drop.
-	parts, _ := listPartitions(context.Background(), db, dbName)
-	if len(parts) < 2 {
-		t.Fatal("need at least 2 partitions to test drop")
-	}
-
-	toDrop := parts[0].Name // drop the first daily partition
-	if err := dropPartitions(context.Background(), db, dbName, []string{toDrop}); err != nil {
-		t.Fatalf("dropPartitions failed: %v", err)
-	}
-
-	// Verify count decreased.
-	partsAfter, _ := listPartitions(context.Background(), db, dbName)
-	if len(partsAfter) != len(parts)-1 {
-		t.Errorf("expected %d partitions after drop, got %d", len(parts)-1, len(partsAfter))
-	}
-
-	// Verify the dropped partition is gone.
-	for _, p := range partsAfter {
-		if p.Name == toDrop {
-			t.Errorf("partition %s should have been dropped", toDrop)
-		}
-	}
-}
-
-// ─── partitionHasData ────────────────────────────────────────────────────────────────
-
-func TestPartitionHasData_empty(t *testing.T) {
-	db, dbName := testutil.CreateTestDB(t)
-	testutil.InitIndexTables(t, db) // creates binlog_events with only p_future
-
-	has, err := partitionHasData(context.Background(), db, dbName)
-	if err != nil {
-		t.Fatalf("partitionHasData failed: %v", err)
-	}
-	if has {
-		t.Error("expected false for empty p_future partition")
-	}
-}
-
-func TestPartitionHasData_withData(t *testing.T) {
-	db, dbName := testutil.CreateTestDB(t)
-	testutil.InitIndexTables(t, db)
-
-	// Insert a row into binlog_events. It will land in p_future since
-	// InitIndexTables creates only p_future.
-	testutil.InsertEvent(t, db,
-		"binlog.000001", 100, 200,
-		time.Now().UTC().Format("2006-01-02 15:04:05"),
-		nil, "testdb", "orders", 1, "1",
-		nil, nil, []byte(`{"id": 1}`),
-	)
-
-	has, err := partitionHasData(context.Background(), db, dbName)
-	if err != nil {
-		t.Fatalf("partitionHasData failed: %v", err)
-	}
-	if !has {
-		t.Error("expected true when p_future has data")
-	}
-}
-
-// ─── addFuturePartitions ─────────────────────────────────────────────────────────────
-
-func TestAddFuturePartitions(t *testing.T) {
-	db, dbName := testutil.CreateTestDB(t)
-	testutil.InitIndexTables(t, db) // only p_future
-
-	startDate := time.Now().UTC().Truncate(time.Hour)
-
-	if err := addFuturePartitions(context.Background(), db, dbName, startDate, 3); err != nil {
-		t.Fatalf("addFuturePartitions failed: %v", err)
-	}
-
-	parts, _ := listPartitions(context.Background(), db, dbName)
-	// Should be 3 hourly + p_future = 4.
-	if len(parts) != 4 {
-		t.Fatalf("expected 4 partitions, got %d", len(parts))
-	}
-
-	// The last one must still be p_future.
-	if parts[len(parts)-1].Name != "p_future" {
-		t.Errorf("expected last partition p_future, got %s", parts[len(parts)-1].Name)
-	}
-
-	// First 3 should be hourly partitions.
-	for i := range 3 {
-		expected := indexer.PartitionName(startDate.Add(time.Duration(i) * time.Hour))
-		if parts[i].Name != expected {
-			t.Errorf("partition %d: expected %s, got %s", i, expected, parts[i].Name)
-		}
-	}
-}
-
 // ─── ArchivePartition ───────────────────────────────────────────────────────────────────
 
 func TestArchivePartition(t *testing.T) {
