+++ cmd/bintrail/rotate_test.go
@@ -1,275 +1,9 @@
 package main
 
 import (
-	"os"
 	"strings"
 	"testing"
-	"time"
-
-	"github.com/dbtrail/bintrail/internal/indexer"
 )
-
-// ─── nextPartitionStart ─────────────────────────────────────────────────────────────
-
-func TestNextPartitionStart_withNamedPartitions(t *testing.T) {
-	partitions := []partitionInfo{
-		{Name: "p_2026021700"},
-		{Name: "p_2026021800"},
-		{Name: "p_2026021900"},
-		{Name: "p_future", Description: "MAXVALUE"},
-	}
-	next := nextPartitionStart(partitions)
-	// Latest is p_2026021900 (hour 00 on 2026-02-19), next should be hour 01.
-	if next.Year() != 2026 || next.Month() != 2 || next.Day() != 19 || next.Hour() != 1 {
-		t.Errorf("expected 2026-02-19 01:00, got %v", next)
-	}
-}
-
-func TestNextPartitionStart_onlyFuture(t *testing.T) {
-	// No named partitions — should return the current hour (UTC, truncated to hour).
-	partitions := []partitionInfo{
-		{Name: "p_future", Description: "MAXVALUE"},
-	}
-	thisHour := time.Now().UTC().Truncate(time.Hour)
-	next := nextPartitionStart(partitions)
-	if !next.Equal(thisHour) {
-		t.Errorf("expected current hour %v, got %v", thisHour, next)
-	}
-}
-
-func TestNextPartitionStart_empty(t *testing.T) {
-	thisHour := time.Now().UTC().Truncate(time.Hour)
-	next := nextPartitionStart(nil)
-	if !next.Equal(thisHour) {
-		t.Errorf("expected current hour %v, got %v", thisHour, next)
-	}
-}
-
-// ─── cutoff logic (hourly retain) ────────────────────────────────────────────
-
-// TestCutoffDropsHourlyPartition verifies that --retain 1h correctly selects
-// the previous-hour partition for dropping. This is the exact scenario from
-// GitHub issue #96: at 00:05 UTC, "p_2026030223" (hour 23) should be eligible
-// for dropping with --retain 1h, because its start time (23:00) is before the
-// cutoff (23:05).
-func TestCutoffDropsHourlyPartition(t *testing.T) {
-	// Simulate: current time = 2026-03-03 00:05:15 UTC, --retain 1h.
-	now := time.Date(2026, 3, 3, 0, 5, 15, 0, time.UTC)
-	retainDur := time.Hour
-	cutoff := now.Add(-retainDur) // 2026-03-02 23:05:15 UTC (no Truncate!)
-
-	partitions := []struct {
-		name     string
-		wantDrop bool
-	}{
-		{"p_2026030222", true},  // hour 22 → 22:00 < 23:05 → drop
-		{"p_2026030223", true},  // hour 23 → 23:00 < 23:05 → drop (the bug was here)
-		{"p_2026030300", false}, // hour 00 → 00:00 is NOT < 23:05 → keep
-		{"p_2026030301", false}, // hour 01 → keep
-	}
-
-	for _, p := range partitions {
-		d, ok := indexer.PartitionDate(p.name)
-		if !ok {
-			t.Fatalf("indexer.PartitionDate(%q) returned false", p.name)
-		}
-		dropped := d.Before(cutoff)
-		if dropped != p.wantDrop {
-			t.Errorf("partition %s (date %v): Before(cutoff %v) = %v, want %v",
-				p.name, d, cutoff, dropped, p.wantDrop)
-		}
-	}
-}
-
-// ─── addFuturePartitions SQL shape ────────────────────────────────────────────
-
-// TestPartitionSpec_shape verifies the generated DDL looks correct without
-// needing a live database.
-func TestPartitionSpec_shape(t *testing.T) {
-	start := time.Date(2026, 2, 20, 14, 0, 0, 0, time.UTC)
-	// Build the partition spec strings the same way addFuturePartitions does.
-	parts := make([]string, 0, 3+1)
-	for i := range 3 {
-		d := start.Add(time.Duration(i) * time.Hour)
-		nextHour := d.Add(time.Hour)
-		parts = append(parts, "PARTITION "+indexer.PartitionName(d)+
-			" VALUES LESS THAN (TO_SECONDS('"+
-			nextHour.UTC().Format("2006-01-02 15:04:05")+"'))")
-	}
-	parts = append(parts, "PARTITION p_future VALUES LESS THAN MAXVALUE")
-
-	if len(parts) != 4 {
-		t.Fatalf("expected 4 parts, got %d", len(parts))
-	}
-	if !strings.Contains(parts[0], "p_2026022014") {
-		t.Errorf("expected p_2026022014 in first spec: %s", parts[0])
-	}
-	if !strings.Contains(parts[0], "2026-02-20 15:00:00") {
-		t.Errorf("expected 2026-02-20 15:00:00 (next hour boundary) in first spec: %s", parts[0])
-	}
-	if !strings.Contains(parts[1], "p_2026022015") {
-		t.Errorf("expected p_2026022015 in second spec: %s", parts[1])
-	}
-	if parts[3] != "PARTITION p_future VALUES LESS THAN MAXVALUE" {
-		t.Errorf("expected p_future catch-all last, got: %s", parts[3])
-	}
-}
-
-// ─── countFuturePartitions ───────────────────────────────────────────────────
-
-func TestCountFuturePartitions(t *testing.T) {
-	ref := time.Date(2026, 2, 19, 12, 0, 0, 0, time.UTC)
-	partitions := []partitionInfo{
-		{Name: "p_2026021910"}, // past
-		{Name: "p_2026021911"}, // past
-		{Name: "p_2026021912"}, // equal to ref → not strictly after
-		{Name: "p_2026021913"}, // future
-		{Name: "p_2026021914"}, // future
-		{Name: "p_future", Description: "MAXVALUE"},
-		{Name: "garbage"},
-	}
-	if got := countFuturePartitions(partitions, ref); got != 2 {
-		t.Errorf("expected 2 future partitions, got %d", got)
-	}
-	if got := countFuturePartitions(nil, ref); got != 0 {
-		t.Errorf("expected 0 for nil slice, got %d", got)
-	}
-}
-
-// ─── autoAdd calculation (declarative --add-future) ─────────────────────────
-
-// TestAutoAddDeclarative verifies the new declarative toAdd formula: --add-future
-// is a target headroom (top-up only), plus auto-replacements for drops.
-func TestAutoAddDeclarative(t *testing.T) {
-	cases := []struct {
-		name        string
-		addFuture   int
-		futureCount int
-		dropped     int
-		noReplace   bool
-		wantToAdd   int
-	}{
-		{"steady state", 1, 1, 0, false, 0},
-		{"cold start", 3, 0, 0, false, 3},
-		{"drops with replace", 1, 1, 1, false, 1},        // 0 top-up + 1 replacement
-		{"drops no replace", 1, 1, 1, true, 0},           // 0 top-up, no replacement
-		{"over target never shrinks", 1, 5, 0, false, 0}, // futureCount > target
-		{"under target top-up", 3, 1, 0, false, 2},
-		{"under target plus drops", 3, 1, 2, false, 4}, // 2 top-up + 2 replacements
-		{"no-replace ignores drops", 3, 1, 2, true, 2}, // 2 top-up only
-		{"no flags", 0, 0, 0, false, 0},
-		{"no-replace over target ignores drops", 1, 5, 3, true, 0},
-	}
-	for _, c := range cases {
-		toAdd := computeToAdd(c.addFuture, c.futureCount, c.dropped, c.noReplace)
-		if toAdd != c.wantToAdd {
-			t.Errorf("%s: expected toAdd=%d, got %d", c.name, c.wantToAdd, toAdd)
-		}
-	}
-}
-
-// TestAutoAddDeclarative_48CycleSimulation replays the daemon rotation loop
-// for 48 cycles to confirm the declarative formula keeps the partition count
-// stable and does not leak (regression test for issue #199).
-func TestAutoAddDeclarative_48CycleSimulation(t *testing.T) {
-	// Start: 12h retention window already populated + 1 future partition,
-	// plus p_future catch-all. Total: 13 named + 1 MAXVALUE = 14.
-	now := time.Date(2026, 4, 1, 0, 0, 0, 0, time.UTC)
-	parts := []partitionInfo{}
-	for i := -11; i <= 1; i++ { // 12 past/current hours + 1 future
-		parts = append(parts, partitionInfo{Name: indexer.PartitionName(now.Add(time.Duration(i) * time.Hour))})
-	}
-	parts = append(parts, partitionInfo{Name: "p_future", Description: "MAXVALUE"})
-
-	retainDur := 12 * time.Hour
-	addFuture := 1
-
-	for cycle := 0; cycle < 48; cycle++ {
-		cycleNow := now.Add(time.Duration(cycle) * time.Hour)
-		cutoff := cycleNow.Add(-retainDur)
-
-		// Drop expired.
-		dropped := 0
-		kept := parts[:0]
-		for _, p := range parts {
-			d, ok := indexer.PartitionDate(p.Name)
-			if ok && d.Before(cutoff) {
-				dropped++
-				continue
-			}
-			kept = append(kept, p)
-		}
-		parts = kept
-
-		// Compute toAdd with the production helper.
-		nowHour := cycleNow.Truncate(time.Hour)
-		futureCount := countFuturePartitions(parts, nowHour)
-		toAdd := computeToAdd(addFuture, futureCount, dropped, false)
-
-		// Append new future partitions after the latest existing named one.
-		startDate := nextPartitionStart(parts)
-		// Strip p_future, re-append later.
-		withoutFuture := parts[:0]
-		for _, p := range parts {
-			if p.Name != "p_future" {
-				withoutFuture = append(withoutFuture, p)
-			}
-		}
-		parts = withoutFuture
-		for i := 0; i < toAdd; i++ {
-			parts = append(parts, partitionInfo{Name: indexer.PartitionName(startDate.Add(time.Duration(i) * time.Hour))})
-		}
-		parts = append(parts, partitionInfo{Name: "p_future", Description: "MAXVALUE"})
-	}
-
-	// Expected stable count: ~15 named partitions + p_future. The important
-	// property is that the count does NOT grow linearly with cycle count
-	// (the pre-fix bug leaked +1 partition per cycle → 48+ after 48 cycles).
-	// Allow a small window around the steady state for boundary transitions.
-	if len(parts) > 17 {
-		t.Errorf("partition count leaked: got %d after 48 cycles, want ≤17 (steady state ~16)", len(parts))
-	}
-	if len(parts) < 14 {
-		t.Errorf("partition count shrank unexpectedly: got %d after 48 cycles, want ≥14", len(parts))
-	}
-
-	// Second 48-cycle run from the same state must return the same count
-	// (true steady state, no drift).
-	startCount := len(parts)
-	for cycle := 48; cycle < 96; cycle++ {
-		cycleNow := now.Add(time.Duration(cycle) * time.Hour)
-		cutoff := cycleNow.Add(-retainDur)
-		dropped := 0
-		kept := parts[:0]
-		for _, p := range parts {
-			d, ok := indexer.PartitionDate(p.Name)
-			if ok && d.Before(cutoff) {
-				dropped++
-				continue
-			}
-			kept = append(kept, p)
-		}
-		parts = kept
-		nowHour := cycleNow.Truncate(time.Hour)
-		futureCount := countFuturePartitions(parts, nowHour)
-		toAdd := computeToAdd(addFuture, futureCount, dropped, false)
-		startDate := nextPartitionStart(parts)
-		withoutFuture := parts[:0]
-		for _, p := range parts {
-			if p.Name != "p_future" {
-				withoutFuture = append(withoutFuture, p)
-			}
-		}
-		parts = withoutFuture
-		for i := 0; i < toAdd; i++ {
-			parts = append(parts, partitionInfo{Name: indexer.PartitionName(startDate.Add(time.Duration(i) * time.Hour))})
-		}
-		parts = append(parts, partitionInfo{Name: "p_future", Description: "MAXVALUE"})
-	}
-	if len(parts) != startCount {
-		t.Errorf("steady-state drift: cycles 48→96 changed count from %d to %d", startCount, len(parts))
-	}
-}
 
 // ─── cobra command wiring ────────────────────────────────────────────────────
 
@@ -315,35 +49,6 @@ func TestRotateCmd_retryDefaultFalse(t *testing.T) {
 	}
 	if f.DefValue != "false" {
 		t.Errorf("expected default retry=false, got %q", f.DefValue)
-	}
-}
-
-// ─── fileExists ──────────────────────────────────────────────────────────────
-
-func TestFileExists(t *testing.T) {
-	dir := t.TempDir()
-
-	// Non-existent file.
-	if fileExists(dir + "/nope.parquet") {
-		t.Error("expected false for non-existent file")
-	}
-
-	// Empty file (0 bytes).
-	empty := dir + "/empty.parquet"
-	if err := os.WriteFile(empty, nil, 0o644); err != nil {
-		t.Fatal(err)
-	}
-	if fileExists(empty) {
-		t.Error("expected false for empty file")
-	}
-
-	// Non-empty file.
-	valid := dir + "/valid.parquet"
-	if err := os.WriteFile(valid, []byte("data"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-	if !fileExists(valid) {
-		t.Error("expected true for non-empty file")
 	}
 }
 
@@ -474,7 +179,7 @@ func TestRunRotate_addFutureAlonePassesFirstGuard(t *testing.T) {
 
 // TestRunRotate_retainAlonePassesFirstGuard verifies that providing only
 // --retain (no --add-future) passes both the "at least one of" guard and
-// the parseRetain check.
+// the retain-parse check.
 func TestRunRotate_retainAlonePassesFirstGuard(t *testing.T) {
 	savedRetain, savedAdd, savedDSN := rotRetain, rotAddFuture, rotIndexDSN
 	t.Cleanup(func() { rotRetain = savedRetain; rotAddFuture = savedAdd; rotIndexDSN = savedDSN })
@@ -488,54 +193,7 @@ func TestRunRotate_retainAlonePassesFirstGuard(t *testing.T) {
 		t.Errorf("first guard should not fire when --retain is set, got: %v", err)
 	}
 	if err != nil && strings.Contains(err.Error(), "--retain:") {
-		t.Errorf("parseRetain should accept '7d', got: %v", err)
-	}
-}
-
-// ─── nextPartitionStart unsorted input ───────────────────────────────────────────
-
-func TestNextPartitionStart_unsortedOrder(t *testing.T) {
-	// Deliberately out of order — must find the maximum date, not the last element.
-	partitions := []partitionInfo{
-		{Name: "p_2026021700"},
-		{Name: "p_2026021900"}, // latest
-		{Name: "p_2026021800"},
-		{Name: "p_future", Description: "MAXVALUE"},
-	}
-	next := nextPartitionStart(partitions)
-	if next.Year() != 2026 || next.Month() != 2 || next.Day() != 19 || next.Hour() != 1 {
-		t.Errorf("expected 2026-02-19 01:00 (max+1h), got %v", next)
-	}
-}
-
-// ─── hiveArchivePath ────────────────────────────────────────────────────────────
-
-func TestHiveArchivePath(t *testing.T) {
-	path, err := hiveArchivePath("/data/archives", "d7bfcfe5-651e-42df-bcab-ffb9149abecb", "p_2026030214")
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	want := "/data/archives/bintrail_id=d7bfcfe5-651e-42df-bcab-ffb9149abecb/event_date=2026-03-02/event_hour=14/events.parquet"
-	if path != want {
-		t.Errorf("hiveArchivePath mismatch\n got: %s\nwant: %s", path, want)
-	}
-}
-
-func TestHiveArchivePath_midnightHour(t *testing.T) {
-	path, err := hiveArchivePath("/data", "abc", "p_2026010100")
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	want := "/data/bintrail_id=abc/event_date=2026-01-01/event_hour=00/events.parquet"
-	if path != want {
-		t.Errorf("hiveArchivePath mismatch\n got: %s\nwant: %s", path, want)
-	}
-}
-
-func TestHiveArchivePath_invalidPartition(t *testing.T) {
-	_, err := hiveArchivePath("/data", "abc", "p_future")
-	if err == nil {
-		t.Fatal("expected error for p_future, got nil")
+		t.Errorf("retain parse should accept '7d', got: %v", err)
 	}
 }
