+++ internal/rotation/loop.go
@@ -1,4 +1,4 @@
-package main
+package rotation
 
 import (
 	"context"
@@ -15,73 +15,68 @@ import (
 	"github.com/dbtrail/bintrail/internal/indexer"
 )
 
-// upRotationSettings is the parsed configuration of `up`'s built-in rotation —
+// Settings is the parsed configuration of `up`'s built-in rotation —
 // the loop that keeps an unattended index from growing until the volume fills
 // and takes the forensic record with it (#420).
-type upRotationSettings struct {
-	enabled   bool
-	retain    time.Duration
-	retainRaw string
-	interval  time.Duration
-	addFuture int
-	// explicit records whether the operator configured --rotate-retain
+type Settings struct {
+	Enabled   bool
+	Retain    time.Duration
+	RetainRaw string
+	Interval  time.Duration
+	AddFuture int
+	// Explicit records whether the operator configured --rotate-retain
 	// themselves (flag or env — bindCommandEnv marks env-set flags Changed).
 	// When false (running on the built-in default), the upgrade guard
 	// refuses to drop pre-existing deep history: an operator who never chose
 	// a retention must not lose months of forensic record to a binary
 	// upgrade.
-	explicit bool
+	Explicit bool
 }
 
-// upRotationCfg carries the parsed settings from runUp's validation to the
-// phase-3 start sites, following the package-global flag style of this file's
-// neighbors (populateStreamFlags et al.).
-var upRotationCfg upRotationSettings
-
-// parseUpRotation validates the --rotate-* flag values and is the sole
-// author of every upRotationSettings field. retain accepts the rotate
+// ParseSettings validates the --rotate-* flag values and is the sole
+// author of every Settings field. retain accepts the rotate
 // command's Nd/Nh forms, plus "off", "0", or "" to disable the built-in
 // rotation entirely. explicit is whether the operator set --rotate-retain
 // themselves (cmd.Flags().Changed — true for flag and env alike).
-func parseUpRotation(retain, interval string, addFuture int, explicit bool) (upRotationSettings, error) {
+func ParseSettings(retain, interval string, addFuture int, explicit bool) (Settings, error) {
 	switch retain {
 	case "off", "0", "":
-		return upRotationSettings{}, nil
+		return Settings{}, nil
 	}
 	dur, err := cliutil.ParseRetain(retain)
 	if err != nil {
-		return upRotationSettings{}, fmt.Errorf("--rotate-retain: %w (or \"off\" to disable)", err)
+		return Settings{}, fmt.Errorf("--rotate-retain: %w (or \"off\" to disable)", err)
 	}
 	iv, err := time.ParseDuration(interval)
 	if err != nil {
-		return upRotationSettings{}, fmt.Errorf("--rotate-interval: %w", err)
+		return Settings{}, fmt.Errorf("--rotate-interval: %w", err)
 	}
 	if iv <= 0 {
-		return upRotationSettings{}, fmt.Errorf("--rotate-interval must be positive, got %q", interval)
+		return Settings{}, fmt.Errorf("--rotate-interval must be positive, got %q", interval)
 	}
 	if addFuture < 0 {
-		return upRotationSettings{}, fmt.Errorf("--rotate-add-future cannot be negative, got %d", addFuture)
+		return Settings{}, fmt.Errorf("--rotate-add-future cannot be negative, got %d", addFuture)
 	}
-	return upRotationSettings{
-		enabled:   true,
-		retain:    dur,
-		retainRaw: retain,
-		interval:  iv,
-		addFuture: addFuture,
-		explicit:  explicit,
+	return Settings{
+		Enabled:   true,
+		Retain:    dur,
+		RetainRaw: retain,
+		Interval:  iv,
+		AddFuture: addFuture,
+		Explicit:  explicit,
 	}, nil
 }
 
-// startUpRotation announces and launches the built-in rotation loop: one
+// StartLoop announces and launches the built-in rotation loop: one
 // cycle immediately, then one every interval, each cycle rotating the boot
 // index database plus every DSN the provider returns (the control plane's
 // per-source databases). Rotation is the secondary job — failures are logged,
 // never fatal to the stream. Returns immediately; the loop stops when ctx is
 // cancelled, and the returned channel closes when it has fully exited (used
 // by tests for deterministic shutdown; production callers may ignore it).
-func startUpRotation(ctx context.Context, s upRotationSettings, dsns func() []string) <-chan struct{} {
+func StartLoop(ctx context.Context, s Settings, dsns func() []string) <-chan struct{} {
 	done := make(chan struct{})
-	if !s.enabled {
+	if !s.Enabled {
 		fmt.Fprintln(os.Stderr, "Built-in rotation: off — the index grows until you rotate it yourself (see `bintrail rotate`)")
 		slog.Info("built-in rotation disabled")
 		close(done)
@@ -90,25 +85,9 @@ func startUpRotation(ctx context.Context, s upRotationSettings, dsns func() []st
 	fmt.Fprintf(os.Stderr,
 		"Built-in rotation: dropping index partitions older than %s every %s, keeping %d future partitions ready.\n"+
 			"  Tune with --rotate-retain / --rotate-interval (or BINTRAIL_ROTATE_RETAIN); disable with --rotate-retain off.\n",
-		s.retainRaw, s.interval, s.addFuture)
+		s.RetainRaw, s.Interval, s.AddFuture)
 	slog.Info("built-in rotation enabled",
-		"retain", s.retainRaw, "interval", s.interval.String(), "add_future", s.addFuture)
-
-	// performRotation reads the rot* package globals (the rotate command's
-	// flag set). Fan them out once before the goroutine starts — same pattern
-	// as populateStreamFlags; nothing else writes them in an `up` process.
-	rotRetain = s.retainRaw
-	rotAddFuture = s.addFuture
-	rotNoReplace = false
-	rotArchiveDir = ""
-	rotArchiveS3 = ""
-	rotBintrailID = ""
-	rotRetry = false
-	// "json" suppresses performRotation's per-partition stdout chatter (it
-	// never emits JSON itself — that is runRotate's wrapper); slog carries
-	// the per-cycle signal instead.
-	rotFormat = "json"
-	rotProtectUnarchived = true
+		"retain", s.RetainRaw, "interval", s.Interval.String(), "add_future", s.AddFuture)
 
 	go func() {
 		defer close(done)
@@ -130,13 +109,13 @@ func startUpRotation(ctx context.Context, s upRotationSettings, dsns func() []st
 					slog.Error("built-in rotation cycle panicked; rotation continues next tick", "panic", r)
 				}
 			}()
-			deferred, failed := runUpRotationCycle(ctx, s, dsns)
+			deferred, failed := runCycle(ctx, s, dsns)
 			if failed || deferred > 0 {
 				unhealthyStreak++
 			} else {
 				unhealthyStreak = 0
 			}
-			if unhealthyStreak >= upRotationEscalateAfter {
+			if unhealthyStreak >= escalateAfter {
 				slog.Error("built-in rotation made no progress for consecutive cycles — the index is growing unbounded (rotation is failing and/or deferring unarchived partitions to a stalled archiving flow; archive the partitions, fix the failure, or set --rotate-retain off and rotate manually)",
 					"consecutive_cycles", unhealthyStreak,
 					"deferred_last_cycle", deferred,
@@ -144,7 +123,7 @@ func startUpRotation(ctx context.Context, s upRotationSettings, dsns func() []st
 			}
 		}
 		cycle()
-		ticker := time.NewTicker(s.interval)
+		ticker := time.NewTicker(s.Interval)
 		defer ticker.Stop()
 		for {
 			select {
@@ -158,16 +137,16 @@ func startUpRotation(ctx context.Context, s upRotationSettings, dsns func() []st
 	return done
 }
 
-// upRotationEscalateAfter is how many consecutive unhealthy cycles (failed
+// escalateAfter is how many consecutive unhealthy cycles (failed
 // or deferring) the loop tolerates before escalating from Warn to Error. A
 // var, not a const, so tests can shrink it.
-var upRotationEscalateAfter = 3
+var escalateAfter = 3
 
-// runUpRotationCycle rotates each index database once. Errors are logged and
+// runCycle rotates each index database once. Errors are logged and
 // the cycle moves to the next DSN — a transient failure self-heals on the
 // next tick. Returns the total number of partitions the protect-unarchived
 // guard deferred this cycle, and whether any database's rotation failed.
-func runUpRotationCycle(ctx context.Context, s upRotationSettings, dsns func() []string) (deferred int, failed bool) {
+func runCycle(ctx context.Context, s Settings, dsns func() []string) (deferred int, failed bool) {
 	for _, dsn := range dedupeDSNs(dsns()) {
 		d, err := rotateOneIndex(ctx, dsn, s)
 		deferred += d
@@ -192,10 +171,28 @@ func dedupeDSNs(in []string) []string {
 	return out
 }
 
-// rotateOneIndex runs one performRotation cycle against a single index DSN,
+// loopOptions builds the Perform Options for one built-in-rotation cycle.
+// Built-in rotation never archives (ArchiveDir empty) — it drops and tops up
+// only — and ALWAYS arms ProtectUnarchived so it can never be the first to
+// destroy data an external archiving flow would preserve; Format "json"
+// suppresses per-partition stdout (slog carries the per-cycle signal). The
+// data-loss safety of `up`'s rotation reduces to this one constructor, so it is
+// kept pure and unit-tested (it replaces the old rot*-global fan-out).
+func loopOptions(retain time.Duration, s Settings) Options {
+	return Options{
+		RetainDur:         retain,
+		RetainRaw:         s.RetainRaw,
+		AddFuture:         s.AddFuture,
+		NoReplace:         false,
+		Format:            "json",
+		ProtectUnarchived: true,
+	}
+}
+
+// rotateOneIndex runs one Perform cycle against a single index DSN,
 // returning the guard-deferred partition count and any failure. Log messages
 // are scrubbed: DSNs (and their passwords) never reach the log.
-func rotateOneIndex(ctx context.Context, dsn string, s upRotationSettings) (int, error) {
+func rotateOneIndex(ctx context.Context, dsn string, s Settings) (int, error) {
 	cfg, err := mysql.ParseDSN(dsn)
 	if err != nil {
 		slog.Warn("built-in rotation: skipping unparseable index DSN",
@@ -214,8 +211,8 @@ func rotateOneIndex(ctx context.Context, dsn string, s upRotationSettings) (int,
 	}
 	defer db.Close()
 
-	retain := s.retain
-	if !s.explicit {
+	retain := s.Retain
+	if !s.Explicit {
 		// Upgrade guard: an operator running on the IMPLICIT default never
 		// chose a retention. If the oldest partition extends far beyond the
 		// default window — the signature of a pre-existing deployment that
@@ -223,7 +220,7 @@ func rotateOneIndex(ctx context.Context, dsn string, s upRotationSettings) (int,
 		// explicit choice. Fresh installs never trip this (they can't
 		// accumulate beyond the window while the loop runs), and a restart
 		// after ordinary downtime stays under the 2× threshold.
-		guarded, oldest, err := upgradeGuardTrips(ctx, db, cfg.DBName, s.retain)
+		guarded, oldest, err := upgradeGuardTrips(ctx, db, cfg.DBName, s.Retain)
 		if err != nil {
 			slog.Warn("built-in rotation: could not evaluate the upgrade guard; skipping drops this cycle",
 				"db", cfg.DBName, "error", config.ScrubDSNText(err.Error(), dsn))
@@ -233,19 +230,19 @@ func rotateOneIndex(ctx context.Context, dsn string, s upRotationSettings) (int,
 			slog.Error("built-in rotation: existing history extends far beyond the default retention — refusing to drop it without an explicit choice",
 				"db", cfg.DBName,
 				"oldest_partition", oldest.UTC().Format("2006-01-02 15:04"),
-				"default_retain", s.retainRaw,
+				"default_retain", s.RetainRaw,
 				"action", "set --rotate-retain explicitly (e.g. 30d to confirm, 90d to keep more, off to disable) or BINTRAIL_ROTATE_RETAIN")
 			retain = 0 // still top up future partitions; no drops
 		}
 	}
 
-	res, err := performRotation(ctx, db, cfg.DBName, retain)
+	res, err := Perform(ctx, db, cfg.DBName, loopOptions(retain, s))
 	if err != nil && ctx.Err() == nil {
 		slog.Warn("built-in rotation cycle failed",
 			"db", cfg.DBName, "error", config.ScrubDSNText(err.Error(), dsn))
-		return res.deferred, err
+		return res.Deferred, err
 	}
-	return res.deferred, nil
+	return res.Deferred, nil
 }
 
 // upgradeGuardTrips reports whether the implicit-default upgrade guard should
